restrict party search for RSVP
I don't want someone to get a list of all possible parties if they just
search for one letter, so let's make sure that we only match the query
string with a space before or after it when loading multiple parties.

Thus searching for "jane" will match both "mary jane" and "jane doe" but
searching for "j" won't match either.